+++ apps/server/src/modules/news/uc/news.uc.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { ObjectId } from '@mikro-orm/mongodb';
-import { LoggerModule } from '@src/core/logger/logger.module';
+import { LoggerModule } from '@src/core/logger';
 import { NotFoundException } from '@nestjs/common/exceptions/not-found.exception';
 import { UnauthorizedException } from '@nestjs/common';
 import { NewsTargetModel, ICreateNews } from '@shared/domain';