+++ apps/server/src/shared/repo/coursegroup/index.ts
@@ -1,2 +1,1 @@
-export * from './course.repo';
 export * from './coursegroup.repo';